+++ Sources/QueueClient/Clients/WorkerDisabler/WorkerDisabler.swift
@@ -1,0 +1,11 @@
+import Dispatch
+import Foundation
+import Models
+
+public protocol WorkerDisabler {
+    func disableWorker(
+        workerId: WorkerId,
+        callbackQueue: DispatchQueue,
+        completion: @escaping (Either<WorkerId, Error>) -> ()
+    )
+}