+++ Sources/RESTMethods/DisableWorker/DisableWorkerPayload.swift
@@ -1,0 +1,12 @@
+import Foundation
+import Models
+
+public final class DisableWorkerPayload: Codable {
+    public let workerId: WorkerId
+    
+    public init(
+        workerId: WorkerId
+    ) {
+        self.workerId = workerId
+    }
+}