+++ Sources/RESTMethods/DisableWorker/WorkerDisabledResponse.swift
@@ -1,0 +1,11 @@
+import DistWorkerModels
+import Foundation
+import Models
+
+public struct WorkerDisabledResponse: Codable, Equatable {
+    public let workerId: WorkerId
+    
+    public init(workerId: WorkerId) {
+        self.workerId = workerId
+    }
+}