+++ Tests/WorkerAlivenessProviderTestHelpers/MutableWorkerAlivenessProvider.swift
@@ -23,6 +23,14 @@ public final class MutableWorkerAlivenessProvider: WorkerAlivenessProvider {
         workerAliveness[workerId] = WorkerAliveness(status: .alive, bucketIdsBeingProcessed: [])
     }
     
+    public func enableWorker(workerId: WorkerId) {
+        workerAliveness[workerId] = WorkerAliveness(status: .alive, bucketIdsBeingProcessed: Set(bucketIdByWorkerId[workerId]))
+    }
+    
+    public func disableWorker(workerId: WorkerId) {
+        workerAliveness[workerId] = WorkerAliveness(status: .disabled, bucketIdsBeingProcessed: Set(bucketIdByWorkerId[workerId]))
+    }
+    
     public var workerAliveness = [WorkerId: WorkerAliveness]()
     
     public func alivenessForWorker(workerId: WorkerId) -> WorkerAliveness {